Added base logic for student data to be exported to Excel

+++ app/src/main/java/com/milesilac/classreadingstats/model/GroupScreeningTest.kt
@@ -4,6 +4,6 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 data class GroupScreeningTest(
-    var score: Int,
+    var score: Double,
     var comprehensionLevel: ComprehensionLevel
 )
