+++ app/src/main/java/com/milesilac/classreadingstats/model/ReadingComprehension.kt
@@ -4,6 +4,6 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 data class ReadingComprehension(
-    var inputPercentage: Float,
+    var inputPercentage: Double,
     var level: LearnerLevel
 )
